# swagger.yaml
openapi: 3.0.0
info:
  title: E-commerece API
  contact: {}
  version: '1.0'
servers:
- url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
  variables: {}
- url: http://localhost:5000/api/v1/orders
  variables: {}
paths:
...
  /users/updateUser:
    patch:
      tags:
      - User
      summary: Update user
      operationId: Updateuser
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateuserRequest'
              - example:
                  password: hello
            example:
              password: hello
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
  /users/updateUserPassword:
    patch:
      tags:
      - User
      summary: update User Password
      operationId: updateUserPassword
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/updateUserPasswordRequest'
              - example:
                  oldPassword: prasish123
                  newPassword: hello123
            example:
              oldPassword: prasish123
              newPassword: hello123
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
  /products:
    post:
      tags:
      - Product
      summary: Create product
      operationId: Createproduct
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateproductRequest'
              - example:
                  name: emperor bed 112
                  price: 23999
                  image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
                  colors:
                  - '#0000ff'
                  - '#000'
                  company: ikea
                  description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
                  category: bedroom
            example:
              name: emperor bed 112
              price: 23999
              image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
              colors:
              - '#0000ff'
              - '#000'
              company: ikea
              description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
              category: bedroom
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
    get:
      tags:
      - Product
      summary: Get all products
      operationId: Getallproducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
  /products/642ea8d96a0d1dd28295b406:
    get:
      tags:
      - Product
      summary: Single products
      operationId: Singleproducts
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
    patch:
      tags:
      - Product
      summary: Update product
      operationId: Updateproduct
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdateproductRequest'
              - example:
                  name: apple
                  price: '250'
                  description: It is a type of food
                  image: ../uploads
                  category: office
                  company: ikea
                  colors: red
                  featured: 'true'
                  freeShipping: 'false'
                  inventory: 15
                  averageRating: 4
            example:
              name: apple
              price: '250'
              description: It is a type of food
              image: ../uploads
              category: office
              company: ikea
              colors: red
              featured: 'true'
              freeShipping: 'false'
              inventory: 15
              averageRating: 4
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
...
  /review:
    post:
      tags:
      - review
      summary: Create review
      operationId: Createreview
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreatereviewRequest'
              - example:
                  name: emperor bed
                  price: 23999
                  image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
                  colors:
                  - '#0000ff'
                  - '#000'
                  company: ikea
                  description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
                  category: bedroom
            example:
              name: emperor bed
              price: 23999
              image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
              colors:
              - '#0000ff'
              - '#000'
              company: ikea
              description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
              category: bedroom
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
    get:
      tags:
      - review
      summary: Get all review
      operationId: Getallreview
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
  /review/123:
    get:
      tags:
      - review
      summary: Get single review
      operationId: Getsinglereview
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
    patch:
      tags:
      - review
      summary: Update review
      operationId: Updatereview
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/UpdatereviewRequest'
              - example:
                  name: emperor bed
                  price: 23999
                  image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
                  colors:
                  - '#0000ff'
                  - '#000'
                  company: ikea
                  description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
                  category: bedroom
            example:
              name: emperor bed
              price: 23999
              image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
              colors:
              - '#0000ff'
              - '#000'
              company: ikea
              description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
              category: bedroom
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
    delete:
      tags:
      - review
      summary: Delete review
      operationId: Deletereview
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      servers:
      - url: https://ecommerce-api-v1-7t56.onrender.com/api/v1/
        variables: {}
...
components:
  schemas:
...
    UpdateuserRequest:
      title: UpdateuserRequest
      required:
      - password
      type: object
      properties:
        password:
          type: string
      example:
        password: hello
    updateUserPasswordRequest:
      title: updateUserPasswordRequest
      required:
      - oldPassword
      - newPassword
      type: object
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
      example:
        oldPassword: prasish123
        newPassword: hello123
    CreateproductRequest:
      title: CreateproductRequest
      required:
      - name
      - price
      - image
      - colors
      - company
      - description
      - category
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          format: int32
        image:
          type: string
        colors:
          type: array
          items:
            type: string
          description: ''
        company:
          type: string
        description:
          type: string
        category:
          type: string
      example:
        name: emperor bed 112
        price: 23999
        image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
        colors:
        - '#0000ff'
        - '#000'
        company: ikea
        description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
        category: bedroom
    UpdateproductRequest:
      title: UpdateproductRequest
      required:
      - name
      - price
      - description
      - image
      - category
      - company
      - colors
      - featured
      - freeShipping
      - inventory
      - averageRating
      type: object
      properties:
        name:
          type: string
        price:
          type: string
        description:
          type: string
        image:
          type: string
        category:
          type: string
        company:
          type: string
        colors:
          type: string
        featured:
          type: string
        freeShipping:
          type: string
        inventory:
          type: integer
          format: int32
        averageRating:
          type: integer
          format: int32
      example:
        name: apple
        price: '250'
        description: It is a type of food
        image: ../uploads
        category: office
        company: ikea
        colors: red
        featured: 'true'
        freeShipping: 'false'
        inventory: 15
        averageRating: 4
    CreatereviewRequest:
      title: CreatereviewRequest
      required:
      - name
      - price
      - image
      - colors
      - company
      - description
      - category
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          format: int32
        image:
          type: string
        colors:
          type: array
          items:
            type: string
          description: ''
        company:
          type: string
        description:
          type: string
        category:
          type: string
      example:
        name: emperor bed
        price: 23999
        image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
        colors:
        - '#0000ff'
        - '#000'
        company: ikea
        description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
        category: bedroom
    UpdatereviewRequest:
      title: UpdatereviewRequest
      required:
      - name
      - price
      - image
      - colors
      - company
      - description
      - category
      type: object
      properties:
        name:
          type: string
        price:
          type: integer
          format: int32
        image:
          type: string
        colors:
          type: array
          items:
            type: string
          description: ''
        company:
          type: string
        description:
          type: string
        category:
          type: string
      example:
        name: emperor bed
        price: 23999
        image: https://dl.airtable.com/.attachmentThumbnails/0446e84c5bca9643de3452a61b2d6195/1b32f48b
        colors:
        - '#0000ff'
        - '#000'
        company: ikea
        description: Cloud bread VHS hell of banjo bicycle rights jianbing umami mumblecore etsy 8-bit pok pok +1 wolf. Vexillologist yr dreamcatcher waistcoat, authentic chillwave trust fund. Viral typewriter fingerstache pinterest pork belly narwhal. Schlitz venmo everyday carry kitsch pitchfork chillwave iPhone taiyaki trust fund hashtag kinfolk microdosing gochujang live-edge
        category: bedroom
...